Fix frequency cap. Switch to long

+++ src-ui/sdr/airspy.cpp
@@ -77,7 +77,7 @@ void SDRAirspy::drawUI()
     ImGui::End();
 }
 
-void SDRAirspy::setFrequency(int frequency)
+void SDRAirspy::setFrequency(long frequency)
 {
     d_frequency = frequency;
     std::memcpy(this->frequency, std::to_string((float)d_frequency / 1e6).c_str(), std::to_string((float)d_frequency / 1e6).length());
