+++ src-ui/sdr/rtlsdr.cpp
@@ -103,7 +103,7 @@ void SDRRtlSdr::drawUI()
     ImGui::End();
 }
 
-void SDRRtlSdr::setFrequency(int frequency)
+void SDRRtlSdr::setFrequency(long frequency)
 {
     d_frequency = frequency;
     std::memcpy(this->frequency, std::to_string((float)d_frequency / 1e6).c_str(), std::to_string((float)d_frequency / 1e6).length());
